Query process image names on Windows
https://trello.com/c/obRmn985

Previously, only PIDs were provided by the WindowsNativeProcessEnumerator.
PIDs are useless alone when choosing a process for inspection. Now, the
running process image name is provided as the description if it can be
queried from the Windows API. If the image name cannot be queried, the user
might want to try again with Administrator privileges.

* Open handle with OpenProcess, requesting QUERY_INFORMATION rights
* Get name with QueryFullProcessImageName in Win32 path format
* Close handle