Find process owner name on Windows
https://trello.com/c/obRmn985

Process owner information was previously not collected on Windows.
Getting the user's name is quite a mess with the Windows API. Lots of
room for refactoring, at least. If the process can't be opened, no owner
data can be read either. User domain name is discarded, as it's
irrelevant for local memory manipulation.
* Open TOKEN_QUERY access token for the process
* Get the TOKEN_USER information for the access token
* Translate the TOKEN_USER PSID to the user's name