Implement memory reading on Linux
https://trello.com/c/rEux0KR6

Reading process memory on Linux is implemented by reading from /proc/$pid/mem.
Accessing other processess' memory on Linux platforms requires users to
run lacuna with sudo rights. This is because access to /proc/$pid/mem is
restricted from other processes.

If `/proc/sys/kernel/yama/ptrace_scope` is set to 0, it is possible to
read another process' memory without sudo by first calling
`ptrace(PTRACE_ATTACH, pid, NULL, NULL)` and then `waitpid(pid, NULL, 0)`
to attach to the target process. This is what, for example, gdb does.

However, ptrace_scope is set to 1 by default, which restricts ptrace to
child processes only. It is less dangerous for users to run lacuna with
sudo than to mess with ptrace_scope, which is a system-wide setting.

Another workaround would be to grant the java binary ptrace capabilities
with `sudo setcap cap_sys_ptrace=eip /path/to/java`, but if the executable
is granted capabilities, the kernel binary loader will require all of java's
dependency shared libraries to reside in trusted paths, or loading them will
fail with "not found". There is a workaround for *that* issue, but now
we're doing workarounds for workarounds so it's time to agree that it's
easier to just have the user `sudo` our program.

It would be possible to actually start the target executable as a child
process in Java, but that's out of scope for this project right now.

Futher reading:
* [_What is the 'ptrace_scope' workaround ..._](http://askubuntu.com/a/153970/537513)
* (About PTRACE_ATTACH failing)[http://stackoverflow.com/a/10163848/996081]
* man ptrace(2)
* [_How to get Oracle java 7 to work with setcap ..._](http://unix.stackexchange.com/q/87978)