DI process opening to WindowsMemoryReader
As part of an effort to abstract away kernel32 from WindowsMemoryReader,
inject a ProcessOpener as a dependency to it. This also fixes leaking
handles in the class. Removed ProcessAccessException, as it was all in all
just a dupe for ProcessOpenException.

Still having to maintain the Kernel32 dependency anyways is suboptimal,
but perhaps we'll come up with something later.

* Inject ProcessOpener to WindowsMemoryReader
* Replace ProcessAccessException with ProcessOpenException
* Make ProcessOpenException a RuntimeException