Major NativeProcessEnumerator changes
After separating pid enumeration and model collection in the
LinuxNativeProcessEnumerator, it became apparent that the implementations
for the Linux and Windows enumerators were the same. As a result, a common
implementation NativeProcessEnumeratorImpl was created in the core
package.

In addition, certain WinApi classes were moved from the windows.winapi
package to the windows package.