Make certain linux classes more testable
Instead of using `Paths.get` inside the linux classes, it's better to
inject `Path` instances and use `Path.resolve` inside the class because
otherwise the file system is coupled to the class. This brings problems
when using virtual file systems during testing because a VFS Path can't be
combined with a real FS path.