Implement memory view in UI
Currently a hex editor view of the process's memory is shown. The editor
is interactive, meaning that all changes are written into the process's
memory. If the backing memory changes, the view is updated if it is
repainted. I think. Still missing go-to-address functionality, so it's a
bit of a scrollfest.