Handle situation where a dead process is chosen
Since the process list is manually refreshed, it is possible for a
process to die before the user chooses the process. Previously, if a
dead process was chosen, a ProcessOpenException was thrown because there
simply were no checks for it. Now, if a ProcessOpenException is thrown,
the current process gets set to `null` and a messagebox is shown.

Closes #16.